Get rid of conditionals in non-code syntax options.
The only conditional was in `char_literals` (a test for EBCDIC).
It can be as well hardcoded in the re2c source code, as there is nothing
language-specific in it.

The next step will be to simplify non-code syntax configurations and
allow only single word values or word lists (no conditionals).