Rework syntax file lexer/parser and the way configurations are stored.
Put syntax configurations in a struct and make it part of the immutable
options instead of storing them in a map. This simplifies their handling
and access, as there's no need to check if a key is in the map, etc.

Merge syntax file lexer/parser into configuration lexer/parser, reuse
existing methods.

Rename list syntax configurations and define them as lists of strings in
quotes (rather than words) so that we can reuse configuration parser for
lists.

Remove separate "features" syntax configurations and replace them with a
single list  of features (to be augmented in future commits).